Move spec files to `src/test`
